:snake: :package: add observability module for service admin
